feat: add auto sleep timer localization strings and update TimePickerDialog.kt
