+++ strings/src/main/res/values-de/strings.xml
@@ -126,7 +126,9 @@
     <string name="sleep_timer.title">Schlaf gut! 🌙</string>
     <string name="sleep_timer.button.increment">Länger</string>
     <string name="sleep_timer.button.decrement">Kürzer</string>
-    <string name="sleep_timer_auto_start">Starte Timer beim Abspielen</string>
+    <string name="auto_sleep_timer">Starte Schlummerfunktion automatisch</string>
+    <string name="auto_sleep_timer_start">Startzeit</string>
+    <string name="auto_sleep_timer_end">Endzeit</string>
     <plurals name="minutes">
         <item quantity="one">%d Minute</item>
         <item quantity="other">%d Minuten</item>
