+++ strings/src/main/res/values/strings.xml
@@ -124,7 +124,9 @@
     <string name="sleep_timer.title">Sweet dreams! 🌙</string>
     <string name="sleep_timer.button.increment">Longer</string>
     <string name="sleep_timer.button.decrement">Shorter</string>
-    <string name="sleep_timer_auto_start">Start timer on playpack</string>
+    <string name="auto_sleep_timer">Start snooze timer on schedule</string>
+    <string name="auto_sleep_timer_start">Start time</string>
+    <string name="auto_sleep_timer_end">End time</string>
     <plurals name="minutes">
         <item quantity="one">%d minute</item>
         <item quantity="other">%d minutes</item>
